refactor: modularize memory plugin embedding pipeline
- split monolithic memory service into focused modules
- add hash/openai/vector providers and shared env utilities
- include integration and write env updates for the new architecture
