+++ packages/memory-plugin/src/embedding/vector-types.ts
@@ -1,0 +1,3 @@
+export interface StoredEmbeddingRow {
+  vector: string;
+}
