+++ packages/memory-plugin/src/index.ts
@@ -1,5 +1,7 @@
 export * from './types.js';
 export * from './service.js';
+export * from './embedding/hash-provider.js';
+export * from './embedding/openai-provider.js';
 export * from './embedding-env.js';
 export * from './write-env.js';
 export * from './integration.js';
